+++ src/pipeline/mod.rs
@@ -2,6 +2,7 @@ use std::any::{type_name, Any};
 use std::collections::HashMap;
 use std::fmt::{Display, Formatter};
 use std::future::Future;
+use std::num::NonZeroUsize;
 use std::panic;
 use std::pin::Pin;
 use std::sync::Arc;
@@ -18,6 +19,88 @@ use sync::Synchronizer;
 mod builder;
 mod io;
 mod sync;
+mod workers;
+
+const DEFAULT_MAX_TASK_COUNT: usize = 100;
+const DEFAULT_READER_BUFFER_SIZE: usize = 30;
+
+/// Options that can be passed to methods in the [PipelineBuilder] when defining stages.
+///
+/// This implements [Default] which makes it easier to specify options when defining stages.
+/// By default, each worker will be allowed 100 concurrent tasks maximum and the buffer of each pipe
+/// is set to 30.
+///
+/// # Examples
+///
+/// ```
+/// use async_pipes::{Pipeline, WorkerOptions};
+///
+/// #[tokio::main]
+/// async fn main() {
+///     let pipeline = Pipeline::builder()
+///         .with_inputs("Pipe", vec![()])
+///         .with_consumer("Pipe", WorkerOptions::default(), |_: ()| async move {
+///             /* ... */
+///         })
+///         .build();
+///
+///     assert!(pipeline.is_ok());
+/// }
+/// ```
+#[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
+pub struct WorkerOptions {
+    /// The maximum number of items allowed per pipe before stages have to wait to write
+    /// more data to the pipe.
+    pub pipe_buffer_size: usize,
+
+    /// The maximum number of tasks that a worker can be concurrently running. Once this number
+    /// is reached in a worker, the worker will poll for tasks completions before spawning more.
+    pub max_task_count: usize,
+}
+
+impl Default for WorkerOptions {
+    fn default() -> Self {
+        Self::default_multi_task()
+    }
+}
+
+impl WorkerOptions {
+    /// Like the [Default] implementation, but specifies `1` for [WorkerOptions::max_task_count].
+    pub fn default_single_task() -> Self {
+        Self {
+            max_task_count: 1,
+            ..Default::default()
+        }
+    }
+
+    /// Specifies `100` for [WorkerOptions::max_task_count] and `30` for
+    /// [WorkerOptions::pipe_buffer_size].
+    pub fn default_multi_task() -> Self {
+        Self {
+            max_task_count: DEFAULT_MAX_TASK_COUNT,
+            pipe_buffer_size: DEFAULT_READER_BUFFER_SIZE,
+        }
+    }
+}
+
+#[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
+struct ValidWorkerOptions {
+    reader_buffer_size: NonZeroUsize,
+    max_task_count: NonZeroUsize,
+}
+
+impl TryFrom<WorkerOptions> for ValidWorkerOptions {
+    type Error = String;
+
+    fn try_from(value: WorkerOptions) -> Result<Self, Self::Error> {
+        Ok(Self {
+            reader_buffer_size: NonZeroUsize::new(value.pipe_buffer_size)
+                .ok_or("reader buffer size must not be zero")?,
+            max_task_count: NonZeroUsize::new(value.max_task_count)
+                .ok_or("max task count must not be zero")?,
+        })
+    }
+}
 
 /// A Box that can hold any value that is [Send].
 ///
@@ -65,49 +148,56 @@ pub type BoxedAnySend = Box<dyn Any + Send + 'static>;
 type ProducerFn = Box<dyn FnMut() -> TaskFuture + Send + 'static>;
 type TaskFn = Box<dyn Fn(BoxedAnySend) -> TaskFuture + Send + Sync + 'static>;
 type IterCastFn = Box<dyn Fn(BoxedAnySend) -> Vec<BoxedAnySend> + Send + Sync + 'static>;
-
 type TaskFuture = Pin<Box<dyn Future<Output = Option<Vec<Option<BoxedAnySend>>>> + Send + 'static>>;
 
-struct ProducerStage {
-    function: ProducerFn,
-    pipes: ProducerPipeNames,
+enum Stage {
+    Producer {
+        function: ProducerFn,
+        pipes: ProducerPipeNames,
+    },
+
+    Regular {
+        function: TaskFn,
+        pipes: TaskPipeNames,
+        options: WorkerOptions,
+    },
+
+    Iterator {
+        stage_type: IterStageType,
+        caster: IterCastFn,
+        pipes: TaskPipeNames,
+        options: WorkerOptions,
+    },
 }
 
+#[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
 struct ProducerPipeNames {
     writers: Vec<String>,
 }
 
-struct TaskStage {
-    function: TaskFn,
-    pipes: TaskPipeNames,
-}
-
+#[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
 struct TaskPipeNames {
     reader: String,
     writers: Vec<String>,
 }
 
-struct IterStage {
-    stage_type: IterStageType,
-    caster: IterCastFn,
-    pipes: TaskPipeNames,
-}
-
+#[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
 enum IterStageType {
     Flatten,
 }
 
+#[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
+struct PipeConfig {
+    name: String,
+    options: ValidWorkerOptions,
+}
+
+#[derive(Debug)]
 struct Pipe<T> {
-    writer: PipeWriter<T>,
     /// Use an option here to "take" it when a reader is used.
     /// Only allow one reader per pipe.
     reader: Option<PipeReader<T>>,
-}
-
-enum Stage {
-    Producer(ProducerStage),
-    Regular(TaskStage),
-    Iterator(IterStage),
+    writer: PipeWriter<T>,
 }
 
 /// Signals sent to stage workers.
@@ -122,7 +212,7 @@ enum StageWorkerSignal {
 impl Display for StageWorkerSignal {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         let signal = match self {
-            Self::Terminate => "SIGTERM",
+            Self::Terminate => "Terminate",
         };
         write!(f, "{signal}")
     }
@@ -143,7 +233,8 @@ impl Display for StageWorkerSignal {
 ///
 /// #[tokio::main]
 /// async fn main() {
-///     let count = Arc::new(Mutex::new(0usize));
+///     use async_pipes::WorkerOptions;
+/// let count = Arc::new(Mutex::new(0usize));
 ///
 ///     let sum = Arc::new(AtomicUsize::new(0));
 ///     let task_sum = sum.clone();
@@ -162,7 +253,7 @@ impl Display for StageWorkerSignal {
 ///                 }
 ///             }
 ///         })
-///         .with_consumer("data", move |value: usize| {
+///         .with_consumer("data", WorkerOptions::default_single_task(), move |value: usize| {
 ///             let sum = task_sum.clone();
 ///             async move {
 ///                 sum.fetch_add(value, SeqCst);
@@ -187,7 +278,8 @@ impl Display for StageWorkerSignal {
 ///
 /// #[tokio::main]
 /// async fn main() {
-///     let count = Arc::new(Mutex::new(0usize));
+///     use async_pipes::WorkerOptions;
+/// let count = Arc::new(Mutex::new(0usize));
 ///
 ///     let odds_sum = Arc::new(AtomicUsize::new(0));
 ///     let task_odds_sum = odds_sum.clone();
@@ -213,13 +305,13 @@ impl Display for StageWorkerSignal {
 ///                 Some(result)
 ///             }
 ///         })
-///         .with_consumer("odds", move |n: usize| {
+///         .with_consumer("odds", WorkerOptions::default_single_task(), move |n: usize| {
 ///             let odds_sum = task_odds_sum.clone();
 ///             async move {
 ///                 odds_sum.fetch_add(n, Ordering::SeqCst);
 ///             }
 ///         })
-///         .with_consumer("evens", move |n: usize| {
+///         .with_consumer("evens", WorkerOptions::default_single_task(), move |n: usize| {
 ///             let evens_sum = task_evens_sum.clone();
 ///             async move {
 ///                 evens_sum.fetch_add(n, Ordering::SeqCst);
@@ -376,23 +468,42 @@ mod tests {
 
     use super::*;
 
-    impl Pipe<BoxedAnySend> {
-        fn new(id: impl Into<String>, has_reader: bool) -> (String, Self) {
-            let id = id.into();
+    macro_rules! pipe_writers {
+        ($count:expr) => {{
+            pipe_writers!($count, ())
+        }};
+
+        ($count:expr, $ch:ty) => {{
+            let sync = Arc::new(Synchronizer::default());
+            let mut writers = Vec::new();
+            let mut rxs = Vec::new();
+            for _ in 0..$count {
+                let id = ulid::Ulid::new().to_string();
+                let (tx, rx) = tokio::sync::mpsc::channel::<$ch>(1);
+                rxs.push(rx);
+                writers.push(PipeWriter::new(format!("{id}"), sync.clone(), tx));
+            }
+            (writers, rxs)
+        }};
+    }
+
+    macro_rules! pipe {
+        ($id:expr, reader=$reader:literal) => {{
+            let id: String = $id.into();
             let sync = Arc::new(Synchronizer::default());
             let (tx, rx) = tokio::sync::mpsc::channel(1);
-            let pipe = Self {
+            let pipe = Pipe {
                 writer: PipeWriter::new(id.clone(), sync.clone(), tx),
-                reader: has_reader.then_some(PipeReader::new(id.clone(), sync, rx)),
+                reader: $reader.then_some(PipeReader::new(id.clone(), sync, rx)),
             };
             (id, pipe)
-        }
+        }};
     }
 
     #[test]
     fn test_find_reader() {
         let pipe_id = "Pipe";
-        let mut pipes = HashMap::from([Pipe::new(pipe_id, true)]);
+        let mut pipes = HashMap::from([pipe!("Pipe", reader = true)]);
 
         let reader = find_reader(pipe_id, &mut pipes);
         assert!(reader.is_ok());
@@ -407,7 +518,7 @@ mod tests {
 
     #[test]
     fn test_find_reader_already_used() {
-        let mut pipes = HashMap::from([Pipe::new("NoReader", false)]);
+        let mut pipes = HashMap::from([pipe!("NoReader", reader = false)]);
 
         let reader = find_reader("NoReader", &mut pipes);
         assert!(reader.is_err());
@@ -417,7 +528,7 @@ mod tests {
     #[test]
     fn test_find_writer() {
         let pipe_id = "Pipe";
-        let pipes = HashMap::from([Pipe::new(pipe_id, true)]);
+        let pipes = HashMap::from([pipe!(pipe_id, reader = true)]);
 
         let writer = find_writer(pipe_id, &pipes);
         assert!(writer.is_ok());
@@ -436,9 +547,9 @@ mod tests {
     #[test]
     fn test_find_writers() {
         let pipes = HashMap::from([
-            Pipe::new("One", true),
-            Pipe::new("Two", false),
-            Pipe::new("Three", true),
+            pipe!("One", reader = true),
+            pipe!("Two", reader = false),
+            pipe!("Three", reader = true),
         ]);
 
         let pipe_ids = vec!["Two".to_string(), "Three".to_string()];
@@ -458,9 +569,9 @@ mod tests {
     #[test]
     fn test_find_writers_open_ended() {
         let pipes = HashMap::from([
-            Pipe::new("One", true),
-            Pipe::new("Two", false),
-            Pipe::new("Three", true),
+            pipe!("One", reader = true),
+            pipe!("Two", reader = false),
+            pipe!("Three", reader = true),
         ]);
 
         let pipe_ids = vec!["Two".to_string(), "Three".to_string(), "Four".to_string()];
@@ -470,13 +581,56 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_write_results() {}
+    async fn test_write_results() {
+        let (writers, mut txs) = pipe_writers!(3, usize);
+        let results = vec![Some(0), None, Some(2)];
 
-    // #[tokio::test]
-    // #[should_panic]
-    // async fn test_write_results_mismatched_lengths() {
-    //     write_results()
-    // }
+        write_results(&writers, results).await;
+
+        assert_eq!(txs.get_mut(0).unwrap().try_recv(), Ok(0));
+        assert!(txs.get_mut(1).unwrap().try_recv().is_err());
+        assert_eq!(txs.get_mut(2).unwrap().try_recv(), Ok(2));
+    }
+
+    #[tokio::test]
+    #[should_panic]
+    async fn test_write_results_panics_on_result_count_mismatch() {
+        let (writers, _txs) = pipe_writers!(5, i32);
+        let results = vec![Some(1), None, None];
+
+        write_results(&writers, results).await;
+    }
+
+    #[test]
+    fn test_downcast_from_pipe() {
+        let value = Box::new(3i8) as BoxedAnySend;
+
+        let casted = downcast_from_pipe::<i8>(value, "some_pipe");
+
+        assert_eq!(casted, Box::new(3i8));
+    }
+
+    #[test]
+    #[should_panic(expected = "failed to downcast input value to i32 from pipe 'some_pipe'")]
+    fn test_downcast_from_pipe_fails() {
+        let value = Box::new(3i8) as BoxedAnySend;
+
+        downcast_from_pipe::<i32>(value, "some_pipe");
+    }
+
+    #[test]
+    fn test_check_join_result_does_nothing_on_ok() {
+        check_join_result(Ok(3usize));
+    }
+
+    #[tokio::test]
+    #[should_panic]
+    async fn test_check_join_result_propagates_panic() {
+        let mut joins = JoinSet::new();
+        joins.spawn(async { panic!("aaaahhhhh") });
+
+        check_join_result(joins.join_next().await.unwrap())
+    }
 
     #[tokio::test]
     async fn test_stage_receives_signal_terminate() {
@@ -484,14 +638,18 @@ mod tests {
 
         let pipeline = Pipeline::builder()
             .with_inputs("pipe", vec![()])
-            .with_consumer("pipe", move |_: ()| {
-                let tx = tx.clone();
-                async move {
-                    tx.send(()).await.unwrap();
-                    tokio::time::sleep(tokio::time::Duration::from_secs(5)).await;
-                    panic!("worker did not terminate!");
-                }
-            })
+            .with_consumer(
+                "pipe",
+                WorkerOptions::default_single_task(),
+                move |_: ()| {
+                    let tx = tx.clone();
+                    async move {
+                        tx.send(()).await.unwrap();
+                        tokio::time::sleep(tokio::time::Duration::from_secs(5)).await;
+                        panic!("worker did not terminate!");
+                    }
+                },
+            )
             .build()
             .unwrap();