+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.3.0 (Jan 13, 2024)
+**Feature(s):**
+* Add `WorkerOptions` struct
+  * Allows specification of max. number of tasks a worker can have and the max. size of a buffer in
+    a pipe.
+* **!! BREAKING !!** - Add parameter to each non-producer stage to pass in worker options.
+  * To resolve the breaking change, for each non-producer stage defined, add in
+    `WorkerOptions::default()` before the task definition/closure.
+
+**Note(s)**:
+* The MSRV is now `1.74.1` (previously `1.63.0`).
+
 ## 0.2.2 (Jan 11, 2024)
 
 **Feature(s):**